Export dialog now only shows PAL entries if working on a PAL project, same for NTSC
svn path=/trunk/kdenlive/; revision=1451

+++ kdenlive/configureprojectdialog.cpp
@@ -71,7 +71,7 @@ void ConfigureProjectDialog::loadTemplates()
     templatesList->clear();
     templatesList->insertStringList(m_app->videoProjectFormats());
     projectFolder->setURL(KdenliveSettings::currentdefaultfolder());
-    templatesList->setCurrentText(m_app->projectFormatName(m_app->projectVideoFormat()));
+    templatesList->setCurrentText(m_app->projectFormatName((uint) m_app->projectVideoFormat()));
     updateDisplay();
 }
